requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via
    #   -r requirements.in
    #   django
    #   django-cors-headers
django==5.1.4
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-debug-toolbar
    #   djangorestframework
django-cors-headers==4.6.0
    # via -r requirements.in
django-debug-toolbar==4.4.6
    # via -r requirements.in
djangorestframework==3.15.2
    # via -r requirements.in
pillow==11.0.0
    # via -r requirements.in
sqlparse==0.5.3
    # via
    #   -r requirements.in
    #   django
    #   django-debug-toolbar
tzdata==2024.2
    # via
    #   -r requirements.in
    #   django
